opal/cma: improve Linux CMA detection
This commit improves the CMA detection when the installed glibc doesn't
have support for CMA. In this case we need to verify that the syscall
numbers in opal/include/opal/sys/cma.h are valid for the architecture.
This verification is done by attempting to use CMA while including the
internal header.

This was back-ported to v2.0.x from master and does not include
support for ARMv8.

Signed-off-by: Nathan Hjelm <hjelmn@me.com>

(cherry picked from commit open-mpi/ompi@4a2bd833026b71064480362bc256f9d6654c18e1)

Signed-off-by: Nathan Hjelm <hjelmn@lanl.gov>